feat: Add endpoint for playing a playlist
This commit adds a new endpoint `/rus_playplaylist/:rusicid` for playing a playlist. The `rus_PlayPlayList` function is implemented to retrieve the songs of the playlist from the database and render them on the `rus_playmusic` template. This feature enhances the application by allowing users to play their created playlists.