Refactor Dockerfiles to update volume paths for Rusic setup
This commit updates the volume paths in the run.sh file to match the new directory structure of the Rusic setup. The paths for the db and thumbnails directories have been modified to "/usr/share/rusicsetup/rusicsetup/db/" and "/usr/share/rusicsetup/rusicsetup/thumbnails/" respectively. This change ensures that the Rusic application can access the correct directories for storing and retrieving data.