Refactor Dockerfiles to update volume paths and add log file
This commit updates the Dockerfiles for RPI/32 and RPI/64 to update the volume paths for the Rusic setup. Additionally, it adds a log file to the Rusic directory. These changes ensure that the Rusic application can access the correct directories for storing and retrieving data, and also have a log file for debugging purposes.