Refactor Dockerfiles to install git package
This commit updates the Dockerfiles for RPI/32 and RPI/64 to include the installation of the git package. This is necessary for certain operations within the Rusic application that require git functionality. By adding the git package to the Dockerfiles, we ensure that the application has the necessary dependencies to function properly.